Replaced floats in math classes with doubles

+++ Game/source/math/quaternion.h
@@ -5,17 +5,17 @@
 class quaternion
 {
 public:
-	float w;
-	float x;
-	float y;
-	float z;
+	double w;
+	double x;
+	double y;
+	double z;
 
 public:
-	quaternion() : w(1.0f), x(0.0f), y(0.0f), z(0.0f) {}
+	quaternion() : w(1.0), x(0.0), y(0.0), z(0.0) {}
 
 	template <typename T>
 	quaternion(const T inW, const T inX, const T inY, const T inZ)
-		: w(static_cast<float>(inW)), x(static_cast<float>(inX)), y(static_cast<float>(inY)), z(static_cast<float>(inZ)) {}
+		: w(static_cast<double>(inW)), x(static_cast<double>(inX)), y(static_cast<double>(inY)), z(static_cast<double>(inZ)) {}
 
 public:
 	vector4d operator*(const vector4d& rhs) const;
